fix(CustomerRepo): sql queries
except the last one

+++ src/main/java/se/experis/task5/models/CustomerSpending.java
@@ -7,9 +7,9 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = false)
 public class CustomerSpending extends Customer {
 
-  private int spendings;
+  private double spendings;
 
-  public CustomerSpending(String firstName, String lastName, String country, String postalCode, String phone, int spendings) {
+  public CustomerSpending(String firstName, String lastName, String country, String postalCode, String phone, double spendings) {
     super(0, firstName, lastName, country, postalCode, phone);
     this.spendings = spendings;
   }
